fix(envelope menu): prevent modals from closing automatically
