+++ package.json
@@ -78,7 +78,6 @@
     "vue": "^2.7.14",
     "vue-autosize": "^1.0.2",
     "vue-click-outside": "^1.1.0",
-    "vue-frag": "^1.4.3",
     "vue-material-design-icons": "^5.2.0",
     "vue-on-click-outside": "^1.0.3",
     "vue-router": "^3.6.5",
